GitButler Integration Commit
This is an integration commit for the virtual branches that GitButler is tracking.

Due to GitButler managing multiple virtual branches, you cannot switch back and
forth between git branches and virtual branches easily. 

If you switch to another branch, GitButler will need to be reinitialized.
If you commit on this branch, GitButler will throw it away.

Here are the branches that are currently applied:
 - develop (refs/gitbutler/develop)
   branch head: 1d2c4ef62e20891722d9c707492466f6364546c4
 - learn-guava (refs/gitbutler/learn-guava)
   branch head: 1d324df9eb8f3ff4af487ffb5125fdea97449153
   - learn-guava/src/main/java/com/kylin/learn/guave/PreconditionsLearn.java
   - learn-guava/src/main/java/com/kylin/learn/guave/ImmutableLearn.java
   - learn-guava/src/main/java/com/kylin/learn/guave/ObjectsLearn.java
   - learn-guava/src/main/java/com/kylin/learn/guave/ThrowableLearn.java
   - learn-guava/src/main/java/com/kylin/learn/guave/OrderingLearn.java

Your previous branch was: refs/heads/learn-guava

The sha for that commit was: 9b730d3182705625922238268931019afe73995f

For more information about what we're doing here, check out our docs:
https://docs.gitbutler.com/features/virtual-branches/integration-branch